# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Bank REST API
  description: REST API для управления банковскими картами, пользователями и транзакциями
  version: 4.7.0
servers:
  - url: http://localhost:8080
    description: Локальный сервер
security:
  - bearerAuth: [ ]
paths:
...
  /api/admin/cards:
    post:
      summary: Создание карты (админ)
      description: Создание новой карты администратором
      operationId: createCard
      tags:
        - Карты (Админ)
      security:
        - bearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CardRequest'
            example:
              userId: 1
              cardNumber: "1234567890123456"
              expiryDate: "2025-12"
              cvv: "123"
              balance: 1000.00
      responses:
        '201':
          description: Карта создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardResponse'
        '400':
          description: Неверные данные запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      summary: Получение списка карт (админ)
      description: Получение списка всех карт
      operationId: getAllCards
      tags:
        - Карты (Админ)
      security:
        - bearerAuth: [ ]
      responses:
        '200':
          description: Список карт
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CardResponse'
  /api/admin/cards/{id}:
    get:
      summary: Получение карты по ID (админ)
      description: Получение информации о карте по её ID
      operationId: getCardById
      tags:
        - Карты (Админ)
      security:
        - bearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Информация о карте
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardResponse'
        '404':
          description: Карта не найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      summary: Обновление карты (админ)
      description: Обновление данных карты администратором
      operationId: updateCard
      tags:
        - Карты (Админ)
      security:
        - bearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CardUpdateRequest'
      responses:
        '200':
          description: Карта обновлена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardResponse'
        '404':
          description: Карта не найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Удаление карты (админ)
      description: Удаление карты по ID
      operationId: deleteCard
      tags:
        - Карты (Админ)
      security:
        - bearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Карта удалена
        '404':
          description: Карта не найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    CardRequest:
      type: object
      properties:
        userId:
          type: integer
          format: int64
          description: ID пользователя
        cardNumber:
          type: string
          description: Номер карты
        expiryDate:
          type: string
          description: Срок действия карты (YYYY-MM)
        cvv:
          type: string
          description: CVV код
        balance:
          type: number
          format: double
          description: Баланс карты
      required:
        - userId
        - cardNumber
        - expiryDate
        - cvv
        - balance
    CardUpdateRequest:
      type: object
      properties:
        cardNumber:
          type: string
          description: Номер карты
        expiryDate:
          type: string
          description: Срок действия карты (YYYY-MM)
        cvv:
          type: string
          description: CVV код
        balance:
          type: number
          format: double
          description: Баланс карты
      required:
        - cardNumber
        - expiryDate
        - cvv
        - balance
    CardResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: ID карты
        userId:
          type: integer
          format: int64
          description: ID пользователя
        cardNumber:
          type: string
          description: Номер карты
        expiryDate:
          type: string
          description: Срок действия карты (YYYY-MM)
        balance:
          type: number
          format: double
          description: Баланс карты
      required:
        - id
        - userId
        - cardNumber
        - expiryDate
        - balance
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          description: Сообщение об ошибке
        status:
          type: integer
          description: HTTP статус код
      required:
        - message
        - status
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT